Removed google dork class, added more dorks, and added boolean to indicate if user wants google dorks in google search.
