+++ wordlists/dorks.txt
@@ -32,3 +32,19 @@ Database_Administration/
 phpSQLiteAdmin/
 0admin/
 super_login%XT%
+/spid/lang/lang.php?lang_path=
+allinurl:.br/index.php?loc=
+PSUser/
+secrets/
+login_admin%XT%
+loginerror/
+superuser%XT%
+authenticate.%XT%
+authentication.%XT%
+auth.%XT%
+authuser.%XT%
+authadmin.%XT%
+cp.%XT%
+modelsearch/login.%XT%
+moderator.%XT%
+moderator/
